Role based settings updated

+++ controller/registration.js
@@ -13,15 +13,20 @@ module.exports.registrationUI = async function(req, res){
         let current_year = year;
         let next_year = year + 1;
         let RN = last.LastRegistration + 1
-        return res.render('./Registration', {ThisRegNumber:RN,past_year, current_year, next_year });
+        return res.render('./Registration', {ThisRegNumber:RN,past_year, current_year, next_year,role:req.user.role });
         
     }else{
-        return res.render('startup')
+        if(req.isAuthenticated && req.user.role === 'Admin'){
+            return res.render('startup',{role:req.user.role})
+        }else{
+            return res.render('Error_403')
+        }
+        
     }
 }
 
 module.exports.getPreview = function(req, res){
-    return res.render('RegistrationPreview', {data:req.body})
+    return res.render('RegistrationPreview', {data:req.body, role:req.user.role})
 }
 
 module.exports.register = async function(req, res){
@@ -30,7 +35,7 @@ module.exports.register = async function(req, res){
         student = await RegisteredStudent.create(req.body);
         lastRegistrationNumber = await AdmissionNo.findOne({});
         RN = lastRegistrationNumber.LastRegistration;
-        await student.updateOne({RegistrationNo:RN+1,SchoolCode:req.user.SchoolCode});
+        await student.updateOne({RegistrationNo:RN+1,SchoolCode:req.user.SchoolCode,RegisteredBy:req.user.full_name});
         await lastRegistrationNumber.updateOne({LastRegistration:+RN+1});
     }catch(err){
         if(student){
@@ -44,17 +49,24 @@ module.exports.register = async function(req, res){
 
 
 module.exports.delete = async function(req, res){
-    try{
-        await RegisteredStudent.findOneAndDelete(req.params.id);
-        return res.status(200).json({
-            message:'Student deleted successfully'
-        })
-    }catch(err){
-        console.log(err);
-        return res.status(404).json({
-            message:'request failed'
+    if(req.user.role === 'Admin'){
+        try{
+            await RegisteredStudent.findOneAndDelete(req.params.id);
+            return res.status(200).json({
+                message:'Student deleted successfully'
+            })
+        }catch(err){
+            console.log(err);
+            return res.status(404).json({
+                message:'request failed'
+            })
+        }
+    }else{
+        return res.status(403).json({
+            message:"Unautorized"
         })
     }
+    
     
 }
 
@@ -69,87 +81,101 @@ function getDate(){
     console.log(currentDate); 
     return currentDate
 }
+
+
 module.exports.admit = async function(req, res){
-    let student,fee, studentData,result_q,result_h,result_f, fee_record;
-    console.log(req.params);
-    try{
-        student = await RegisteredStudent.findOne({RegistrationNo:req.params.id, SchoolCode:req.user.SchoolCode});
-        studentData = JSON.parse(JSON.stringify(student));
-        delete studentData._id;
-        //await RegisterdStudent.updateOne({student}, {$unset : {_id:1}});
-        //student.delete('_id');
-        let newStudent = await Student.create(studentData);
-        let adm = await AdmissionNo.findOne({SchoolCode:req.user.SchoolCode});
-        let lastAdmissionNo = +adm.LastAdmission;
-        await newStudent.update({AdmissionNo: lastAdmissionNo+1, SchoolCode:req.user.SchoolCode});
-        await adm.update({LastAdmission:lastAdmissionNo+1});
-        adm.save();
-        fee = await FeeStructure.findOne({Class:studentData.Class,SchoolCode:req.user.SchoolCode});
-        fee_record = await FeeSchema.create({
-            AdmissionNo:lastAdmissionNo+1,
-            Class:studentData.Class,
-            Total:fee.Fees,
-            Remaining: fee.Fees,
-            Paid:0,
-            SchoolCode:req.user.SchoolCode
-        });
+    if(req.user.role === 'Admin'){
+        let student,fee, studentData,result_q,result_h,result_f, fee_record;
+        console.log(req.params);
+        try{
+            student = await RegisteredStudent.findOne({RegistrationNo:req.params.id, SchoolCode:req.user.SchoolCode});
+            studentData = JSON.parse(JSON.stringify(student));
+            delete studentData._id;
+            //await RegisterdStudent.updateOne({student}, {$unset : {_id:1}});
+            //student.delete('_id');
+            let newStudent = await Student.create(studentData);
+            let adm = await AdmissionNo.findOne({SchoolCode:req.user.SchoolCode});
+            let lastAdmissionNo = +adm.LastAdmission;
+            await newStudent.update({AdmissionNo: lastAdmissionNo+1, SchoolCode:req.user.SchoolCode});
+            await adm.update({LastAdmission:lastAdmissionNo+1});
+            adm.save();
+            fee = await FeeStructure.findOne({Class:studentData.Class,SchoolCode:req.user.SchoolCode});
+            fee_record = await FeeSchema.create({
+                AdmissionNo:lastAdmissionNo+1,
+                Class:studentData.Class,
+                Total:fee.Fees,
+                Remaining: fee.Fees,
+                Paid:0,
+                SchoolCode:req.user.SchoolCode
+            });
 
-        result_q = await Result.create({
-            AdmissionNo: lastAdmissionNo+1,
-            Class:studentData.Class,
-            Term: 'Quarterly',
-            SchoolCode:req.user.SchoolCode
-        });
+            result_q = await Result.create({
+                AdmissionNo: lastAdmissionNo+1,
+                Class:studentData.Class,
+                Term: 'Quarterly',
+                SchoolCode:req.user.SchoolCode
+            });
 
-        result_h = await Result.create({
-            AdmissionNo: lastAdmissionNo+1,
-            Class:studentData.Class,
-            Term: 'Half-Yearly',
-            SchoolCode:req.user.SchoolCode
-        });
+            result_h = await Result.create({
+                AdmissionNo: lastAdmissionNo+1,
+                Class:studentData.Class,
+                Term: 'Half-Yearly',
+                SchoolCode:req.user.SchoolCode
+            });
 
-        result_f = await Result.create({
-            AdmissionNo: lastAdmissionNo+1,
-            Class:studentData.Class,
-            Term: 'Final',
-            SchoolCode:req.user.SchoolCode
-        });
-        await RegisteredStudent.deleteOne(student);
-        await TCRecords.create({
-            AdmissionNo:lastAdmissionNo+1,
-            AdmissionClass: studentData.Class,
-            AdmissionDate: getDate(),
-            SchoolCode:req.user.SchoolCode
-        })
-        return res.status(200).json({
-            message:'Student admitted'
-        });
+            result_f = await Result.create({
+                AdmissionNo: lastAdmissionNo+1,
+                Class:studentData.Class,
+                Term: 'Final',
+                SchoolCode:req.user.SchoolCode
+            });
+            await RegisteredStudent.deleteOne(student);
+            await TCRecords.create({
+                AdmissionNo:lastAdmissionNo+1,
+                AdmissionClass: studentData.Class,
+                AdmissionDate: getDate(),
+                SchoolCode:req.user.SchoolCode
+            })
+            return res.status(200).json({
+                message:'Student admitted'
+            });
 
-    }catch(err){
-        if(student){
-            Student.remove(student)
+        }catch(err){
+            if(student){
+                Student.remove(student)
+            }
+            if(fee_record){
+                FeeStructure.remove(fee_record)
+            }
+            if(result_q){
+                Result.remove(result_q);
+            }
+            if(result_h){
+                Result.remove(result_h);
+            }
+            if(result_f){
+                Result.remove(result_f);
+            }
+            console.log(err);
+            return res.status(503).json({
+                message:'Unable to admit, please try again later'
+            })
         }
-        if(fee_record){
-            FeeStructure.remove(fee_record)
-        }
-        if(result_q){
-            Result.remove(result_q);
-        }
-        if(result_h){
-            Result.remove(result_h);
-        }
-        if(result_f){
-            Result.remove(result_f);
-        }
-        console.log(err);
-        return res.status(503).json({
-            message:'Unable to admit, please try again later'
+    }else{
+        return res.status(403).json({
+            message:"Unautorized"
         })
     }
+    
 }
 
 
 module.exports.download = async function(req, res){
-    let data = await RegisteredStudent.findOne({RegistrationNo:req.params.id});
-    return res.render('RegistrationForm', {data});
+    if(req.user.role === 'Admin'){
+        let data = await RegisteredStudent.findOne({RegistrationNo:req.params.id});
+        return res.render('RegistrationForm', {data, role:req.user.role});
+    }else{
+        return res.render('Error_403')
+    }
+    
 }
